PPP-31 & PPP-33 feature filter implemented, folder structure reorganized and dog form hard-coded dog breed list replaced
