+++ src/components/Filters/DogFilterStyles.js
@@ -2,12 +2,11 @@ import styled from "styled-components";
 
 const WrapDiv = styled.div`
   position: fixed;
-  top: 12rem;
-  left: 20px;
+  top: 20rem;
+  left: 10px;
   display: block;
   text-decoration: none;
-  color: white;
-  padding: 10px 30px;
+  padding: 5px 10px;
   width: 100%;
   max-width: 14em;
   text-align: left;
@@ -15,11 +14,12 @@ const WrapDiv = styled.div`
     font-weight: bold;
     color: white;
     margin-left: 5px;
-    margin-top: 20px;
-    margin-bottom: 10px;
+    margin-top: 40px;
+    margin-bottom: 15px;
   }
   @media (max-width: 1220px) {
-    top: 19rem;
+    top: 25rem;
+    left: 2px;
     font-size: smaller;
   }
 `;
