+++ src/components/Filters/SitterFilterStyles.js
@@ -2,11 +2,12 @@ import styled from "styled-components";
 
 const WrapDiv = styled.div`
   position: fixed;
-  top: 20rem;
-  left: 10px;
+  top: 12rem;
+  left: 15px;
   display: block;
   text-decoration: none;
-  padding: 5px 10px;
+  color: white;
+  padding: 10px 30px;
   width: 100%;
   max-width: 14em;
   text-align: left;
@@ -14,12 +15,11 @@ const WrapDiv = styled.div`
     font-weight: bold;
     color: white;
     margin-left: 5px;
-    margin-top: 20px;
+    margin-top: 30px;
     margin-bottom: 10px;
   }
   @media (max-width: 1220px) {
-    top: 25rem;
-    left: 2px;
+    top: 19rem;
     font-size: smaller;
   }
 `;
