+++ src/pages/HeaderPages/Home.js
@@ -1,4 +1,4 @@
-import ImageHeader from "../../components/UI/ImageHeader";
+import HeaderWithImage from "../../components/UI/HeaderWithImage";
 import IconWrapper from "../../components/HomePageComponents/Wrappers/IconWrapper";
 import ButtonWrapper from "../../components/HomePageComponents/Wrappers/ButtonWrapper";
 import DogIcon from "../../components/HomePageComponents/Buttons/DogButton";
@@ -7,7 +7,7 @@ import SitterIcon from "../../components/HomePageComponents/Buttons/SitterButton
 const Home = () => {
   return (
     <>
-      <ImageHeader />
+      <HeaderWithImage />
       <IconWrapper />
       <ButtonWrapper>
         <SitterIcon />
